Refactor: Move core and widget classes to subdirectories
