+++ lib/balloon.dart
@@ -1,0 +1,10 @@
+library balloon_widget;
+
+export 'src/core/balloon_nip_position.dart';
+export 'src/core/balloon_shadow.dart';
+export 'src/core/balloon_painter.dart';
+export 'src/core/balloon_clipper.dart';
+
+export 'src/widgets/balloon_tap_delegator.dart';
+export 'src/widgets/positioned_balloon.dart';
+export 'src/widgets/balloon_widget.dart';
