+++ pubspec.yaml
@@ -11,6 +11,7 @@ dependencies:
     sdk: flutter
 
   defer_pointer: ^0.0.2
+  meta: any
 
 dev_dependencies:
   flutter_test:
